160222.21:37
Сделал все синглотоны глобальными обьектами путём DontDestroyOnLoad()
Добавил скрипт для глобальных обьектов
Добавил StateBank который после определенного количества вызовов вызывает действие (пока не проверял работоспособность)

+++ Assets/Скрипты/Сохранения/SaveLoad.cs
@@ -25,10 +25,7 @@ public static class SaveLoad
 
     public static void DeleteSaves()
     {
-        try
-        {
+        if (File.Exists(Application.persistentDataPath + "/save.json"))
             File.Delete(Application.persistentDataPath + "/save.json");
-        }
-        catch { }
     }
 }
